requirements.txt
# ── LLM inference (Apple Silicon MLX) ────────────────
mlx-lm>=0.21.0

# ── LangGraph / LangChain ────────────────────────────
langgraph>=0.2.0
langchain-core>=0.3.0

# ── Vector Memory (Long-term knowledge) ─────────────
chromadb>=0.5.0
sentence-transformers>=3.0.0

# ── Web scraping & search ────────────────────────────
ddgs>=9.0.0
trafilatura>=2.0.0
lxml_html_clean>=0.4.0
beautifulsoup4>=4.12.0
requests>=2.31.0
feedparser>=6.0.0
playwright>=1.58.0

# ── Telegram ─────────────────────────────────────────
python-telegram-bot>=21.0
telethon>=1.36.0

# ── Notion ───────────────────────────────────────────
notion-client>=2.2.0

# ── Utilities ────────────────────────────────────────
python-dotenv>=1.0.0
